refactor(pattern): regroup dictionaries by language, lower scores
Dictionaries are now grouped by scope rather than thrown into
one `world/` bucket:

- `assets/dictionaries/world/` — language-agnostic brand-name
  lists. Now contains only `cryptocurrencies` (BTC, Ethereum,
  Tether, … — the same strings in every language).
- `assets/dictionaries/en/` — terms written in English that
  translate when the document language changes:
  - `finance/currencies` (USD, US Dollar, …)
  - `personal/languages` (English, en, French, fr, …)
  - `personal/nationalities` (American, French, Japanese, …)
  - `personal/religions` (Christian, Muslim, Buddhist, …)

Each English dictionary TOML declares `languages = ["en"]`, so
the recognizer filters them out at runtime when the caller
asserts a non-English document via `RecognizerInput.language`.
Future Spanish, French, German variants land at
`assets/dictionaries/<lang>/` without touching the runtime.

Scores: lowered the dictionary baseline from `0.85` to `0.4`
(`[0.85, 0.30, 0.85]` → `[0.4, 0.2, 0.4]` on the multi-column
languages list) to match Microsoft Presidio's typical
deny-list-recognizer baseline. At 0.85 every "American" or
"Christian" mention in any English document flooded results;
at 0.4 the context-enhancer pipeline lifts them when nearby
keywords like "nationality" / "religion" / "currency" fire.

`shipped::dictionaries::{en,world}` module re-shuffled to
match the asset tree.

Dropped count-asserting tests on the `all()` functions in
both `shipped::patterns` and `shipped::dictionaries` — they
break every time a new pattern or dictionary ships without
catching any real bug; replaced with a single
`every_shipped_pattern_has_variants` / `every_shipped_dictionary_parses`
loop.

Asset docs:

- Added `assets/README.md`: walks the directory tree, lists
  the score-tier conventions, names the built-in validators,
  points to PRESIDIO.md.
- Renamed `assets/NOTICE.md` to `assets/PRESIDIO.md` and
  retitled it from "third-party attribution" to "Presidio
  attribution" — the file always was just the upstream MIT
  license requirement; the new name says so.
