+++ crates/nvisy-pattern/assets/dictionaries/en/finance/currencies.csv
