+++ crates/nvisy-pattern/assets/dictionaries/en/finance/currencies.toml
@@ -1,3 +1,4 @@
 name = "currencies"
 label = "currency"
-score = 0.85
+languages = ["en"]
+score = 0.4
