+++ crates/nvisy-pattern/assets/dictionaries/en/personal/languages.csv
