+++ crates/nvisy-pattern/assets/dictionaries/en/personal/nationalities.toml
@@ -1,3 +1,4 @@
 name = "nationalities"
 label = "nationality"
-score = 0.85
+languages = ["en"]
+score = 0.4
