+++ crates/nvisy-pattern/assets/dictionaries/en/personal/nationalities.txt
