+++ crates/nvisy-pattern/assets/dictionaries/en/personal/religions.toml
@@ -1,3 +1,4 @@
 name = "religions"
 label = "religion"
-score = 0.85
+languages = ["en"]
+score = 0.4
