+++ crates/nvisy-pattern/assets/dictionaries/en/personal/religions.txt
