+++ crates/nvisy-pattern/assets/dictionaries/world/finance/cryptocurrencies.toml
@@ -1,3 +1,3 @@
 name = "cryptocurrencies"
 label = "currency"
-score = 0.85
+score = 0.4
