+++ crates/nvisy-pattern/src/shipped/dictionaries/world.rs
@@ -1,4 +1,9 @@
-//! Universal dictionaries — apply regardless of jurisdiction.
+//! Universal dictionaries — apply regardless of jurisdiction or
+//! language.
+//!
+//! These are brand-name lists whose terms transcend locale.
+//! Cryptocurrency names (`Bitcoin`, `Ethereum`) and tickers
+//! (`BTC`, `ETH`) are the same string in every language.
 
 use crate::{__shipped_dictionary as shipped_dictionary, Dictionary};
 
@@ -9,41 +14,9 @@ shipped_dictionary!(
         from "../../../assets/dictionaries/world/finance/cryptocurrencies.toml"
         with csv "../../../assets/dictionaries/world/finance/cryptocurrencies.csv"
 );
-shipped_dictionary!(
-    /// Fiat currency names and ISO 4217 codes (USD, US Dollar,
-    /// EUR, Euro, …).
-    fn currencies
-        from "../../../assets/dictionaries/world/finance/currencies.toml"
-        with csv "../../../assets/dictionaries/world/finance/currencies.csv"
-);
-shipped_dictionary!(
-    /// Human-language names and ISO 639 codes (English, en,
-    /// French, fr, …).
-    fn languages
-        from "../../../assets/dictionaries/world/personal/languages.toml"
-        with csv "../../../assets/dictionaries/world/personal/languages.csv"
-);
-shipped_dictionary!(
-    /// Demonyms and nationality terms (American, French, …).
-    fn nationalities
-        from "../../../assets/dictionaries/world/personal/nationalities.toml"
-        with text "../../../assets/dictionaries/world/personal/nationalities.txt"
-);
-shipped_dictionary!(
-    /// Religious affiliations (Christianity, Islam, …).
-    fn religions
-        from "../../../assets/dictionaries/world/personal/religions.toml"
-        with text "../../../assets/dictionaries/world/personal/religions.txt"
-);
 
 /// Every world-scoped built-in dictionary.
 #[must_use]
 pub fn all() -> Vec<Dictionary> {
-    vec![
-        cryptocurrencies(),
-        currencies(),
-        languages(),
-        nationalities(),
-        religions(),
-    ]
+    vec![cryptocurrencies()]
 }
